organize the whole web application
