--- public/index.html
@@ -1,63 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <title>Mosquito Dashboard</title>
-    <script type="module" src="/javascript/app.js"></script>
-    <link rel="stylesheet" href="/stylesheets/app.css">
-    <link rel="stylesheet" href="/stylesheets/overseer.css">
-  </head>
-  <body>
-    <h1>Mosquito Dashboard</h1>
-
-    <div id="tabs">
-      <a href="" data-tab-target="queues" class="active">Queues</a>
-      <a href="" data-tab-target="overseers">Overseers</a>
-      <a href="" data-tab-target="messages">Messages</a>
-    </div>
-
-    <div id="tab-content">
-      <div id="queues" class="active">
-        <mosquito-queue-list></mosquito-queue-list>
-      </div>
-      <div id="messages">
-        <mosquito-message-stream></mosquito-message-stream>
-      </div>
-      <div id="overseers">
-      </div>
-    </div>
-
-    <template id="overseer">
-      <div class="overseer">
-        <h3>
-          <span class="overseer-id"></span>
-          <span class="summary" class="hidden"></span>
-        </h3>
-
-        <div class="last-active-at" class="hidden"></div>
-
-        <table class="executors">
-          <thead>
-            <tr>
-              <th>ID</th>
-              <th>Current Job</th>
-            </tr>
-          </thead>
-          <tbody></tbody>
-        </table>
-      </div>
-    </template>
-
-    <template id="executor">
-      <tr class="executor-row">
-        <td class="executor-id"></td>
-        <td class="working-on"></td>
-      </tr>
-
-      <tr class="progress-row">
-        <td colspan="3">
-          <div class="progress-bar"></div>
-        </td>
-      </tr>
-    </template>
-  </body>
-</html>
