+++ public/stylesheets/dashboard.css
@@ -1,0 +1,18 @@
+mosquito-message-stream {
+  height: 56vh;
+}
+
+#tab-content {
+  display: flex;
+  border: 1px solid #dddddd;
+}
+
+#tab-content > * {
+  display: none;
+  width: 100%;
+  margin: 1rem;
+}
+
+#tab-content .active {
+  display: block;
+}
