+++ src/http/root.cr
@@ -1,0 +1,3 @@
+get "/" do |env|
+  env.redirect "/overseers"
+end
