+++ src/views/layout.html.ecr
@@ -1,0 +1,35 @@
+<!doctype html>
+<html>
+<head>
+  <title>Mosquito</title>
+  <link rel="stylesheet" type="text/css" href="/stylesheets/app.css">
+  <script type="module" src="/javascript/app.js"></script>
+
+  <%= yield_content "header" %>
+  <% if Kemal.config.env == "development" %>
+    <script type="module" src="/javascript/hot_reload.js"></script>
+  <% end %>
+</head>
+<body>
+  <!--div id="queues" class="active">
+    <mosquito-queue-list></mosquito-queue-list>
+  </div>
+  <div id="messages">
+    <mosquito-message-stream></mosquito-message-stream>
+  </div-->
+
+
+  <h1>Mosquito Dashboard (<%= Current.tab %>)</h1>
+
+  <div id="tabs">
+    <a href="/queues"  class="current-tab-<%= Current.tab == :queues %>">Queues</a>
+    <a href="/overseers" class="current-tab-<%= Current.tab == :overseers %>">Overseers</a>
+    <a href="/messages"  class="current-tab-<%= Current.tab == :messages %>">Messages</a>
+    <%= yield_content "extra_tabs" %>
+  </div>
+
+  <div id="content">
+    <%= content %>
+  </div>
+</body>
+</html>
